Migration from Dagger2 to Hilt for GitHub Browser sample.
