+++ GithubBrowserSample/app/src/main/java/com/android/example/github/GithubApp.kt
@@ -16,26 +16,17 @@
 
 package com.android.example.github
 
-import android.app.Activity
 import android.app.Application
-import com.android.example.github.di.AppInjector
-import dagger.android.DispatchingAndroidInjector
-import dagger.android.HasActivityInjector
+import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
-import javax.inject.Inject
 
-
-class GithubApp : Application(), HasActivityInjector {
-    @Inject
-    lateinit var dispatchingAndroidInjector: DispatchingAndroidInjector<Activity>
+@HiltAndroidApp
+class GithubApp : Application() {
 
     override fun onCreate() {
         super.onCreate()
         if (BuildConfig.DEBUG) {
             Timber.plant(Timber.DebugTree())
         }
-        AppInjector.init(this)
     }
-
-    override fun activityInjector() = dispatchingAndroidInjector
 }
