+++ GithubBrowserSample/app/src/main/java/com/android/example/github/di/NetworkModule.kt
@@ -1,0 +1,27 @@
+package com.android.example.github.di
+
+import com.android.example.github.api.GithubService
+import com.android.example.github.util.LiveDataCallAdapterFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(ApplicationComponent::class)
+object NetworkModule {
+
+    @Singleton
+    @Provides
+    fun provideGithubService(): GithubService {
+        return Retrofit.Builder()
+                .baseUrl("https://api.github.com/")
+                .addConverterFactory(GsonConverterFactory.create())
+                .addCallAdapterFactory(LiveDataCallAdapterFactory())
+                .build()
+                .create(GithubService::class.java)
+    }
+}
