+++ GithubBrowserSample/app/src/main/java/com/android/example/github/di/PersistenceModule.kt
@@ -1,0 +1,38 @@
+package com.android.example.github.di
+
+import android.app.Application
+import androidx.room.Room
+import com.android.example.github.db.GithubDb
+import com.android.example.github.db.RepoDao
+import com.android.example.github.db.UserDao
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(ApplicationComponent::class)
+object PersistenceModule {
+
+    @Singleton
+    @Provides
+    fun provideDb(app: Application): GithubDb {
+        return Room
+                .databaseBuilder(app, GithubDb::class.java, "github.db")
+                .fallbackToDestructiveMigration()
+                .build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideUserDao(db: GithubDb): UserDao {
+        return db.userDao()
+    }
+
+    @Singleton
+    @Provides
+    fun provideRepoDao(db: GithubDb): RepoDao {
+        return db.repoDao()
+    }
+}
