+++ GithubBrowserSample/app/src/main/java/com/android/example/github/di/RepositoryModule.kt
@@ -1,0 +1,11 @@
+package com.android.example.github.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+
+@Module
+@InstallIn(ApplicationComponent::class)
+object RepositoryModule {
+
+}
