+++ GithubBrowserSample/app/src/main/java/com/android/example/github/ui/repo/RepoFragment.kt
@@ -26,7 +26,6 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.FragmentNavigatorExtras
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
@@ -35,22 +34,19 @@ import com.android.example.github.AppExecutors
 import com.android.example.github.R
 import com.android.example.github.binding.FragmentDataBindingComponent
 import com.android.example.github.databinding.RepoFragmentBinding
-import com.android.example.github.di.Injectable
 import com.android.example.github.ui.common.RetryCallback
 import com.android.example.github.util.autoCleared
+import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 /**
  * The UI Controller for displaying a Github Repo's information with its contributors.
  */
-class RepoFragment : Fragment(), Injectable {
 
-    @Inject
-    lateinit var viewModelFactory: ViewModelProvider.Factory
+@AndroidEntryPoint
+class RepoFragment : Fragment() {
 
-    val repoViewModel: RepoViewModel by viewModels {
-        viewModelFactory
-    }
+    private val repoViewModel: RepoViewModel by viewModels()
 
     @Inject
     lateinit var appExecutors: AppExecutors
@@ -75,14 +71,14 @@ class RepoFragment : Fragment(), Injectable {
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+            inflater: LayoutInflater, container: ViewGroup?,
+            savedInstanceState: Bundle?
     ): View? {
         val dataBinding = DataBindingUtil.inflate<RepoFragmentBinding>(
-            inflater,
-            R.layout.repo_fragment,
-            container,
-            false
+                inflater,
+                R.layout.repo_fragment,
+                container,
+                false
         )
         dataBinding.retryCallback = object : RetryCallback {
             override fun retry() {
