+++ GithubBrowserSample/app/src/main/java/com/android/example/github/ui/user/UserViewModel.kt
@@ -16,6 +16,7 @@
 
 package com.android.example.github.ui.user
 
+import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -27,11 +28,9 @@ import com.android.example.github.util.AbsentLiveData
 import com.android.example.github.vo.Repo
 import com.android.example.github.vo.Resource
 import com.android.example.github.vo.User
-import javax.inject.Inject
 
 @OpenForTesting
-class UserViewModel
-@Inject constructor(userRepository: UserRepository, repoRepository: RepoRepository) : ViewModel() {
+class UserViewModel @ViewModelInject constructor(userRepository: UserRepository, repoRepository: RepoRepository) : ViewModel() {
     private val _login = MutableLiveData<String?>()
     val login: LiveData<String?>
         get() = _login
