+++ GithubBrowserSample/versions.gradle
@@ -22,7 +22,7 @@
 ext.deps = [:]
 def versions = [:]
 versions.activity = '1.1.0'
-versions.android_gradle_plugin = '4.0.0'
+versions.android_gradle_plugin = '4.1.1'
 versions.annotations = "1.0.0"
 versions.apache_commons = "2.5"
 versions.appcompat = "1.2.0-alpha02"
@@ -63,6 +63,8 @@ versions.timber = "4.7.1"
 versions.transition = "1.3.0"
 versions.truth = "1.0.1"
 versions.work = "2.2.0"
+versions.hilt = "2.28-alpha"
+versions.androidx_hilt = "1.0.0-alpha02"
 ext.versions = versions
 
 def build_versions = [:]
@@ -215,6 +217,14 @@ work.testing = "androidx.work:work-testing:$versions.work"
 work.firebase = "androidx.work:work-firebase:$versions.work"
 work.runtime_ktx = "androidx.work:work-runtime-ktx:$versions.work"
 deps.work = work
+
+def hilt = [:]
+hilt.runtime = "com.google.dagger:hilt-android:$versions.hilt"
+hilt.plugin = "com.google.dagger:hilt-android-gradle-plugin:$versions.hilt"
+hilt.compiler = "com.google.dagger:hilt-android-compiler:$versions.hilt"
+hilt.lifecycle_viewmodel = "androidx.hilt:hilt-lifecycle-viewmodel:$versions.androidx_hilt"
+hilt.androidx_compiler = "androidx.hilt:hilt-compiler:$versions.androidx_hilt"
+deps.hilt = hilt
 
 ext.deps = deps
